Remove classic refresh from Openstack InfraManager
