+++ config/settings.yml
@@ -7,10 +7,8 @@
     :parallel_thread_limit: 0
   :openstack_network:
     :is_admin: false
-    :inventory_object_refresh: true
   :openstack_infra:
     :is_admin: false
-    :inventory_object_refresh: true
   :cinder:
     :is_admin: false
 :ems:
