+++ spec/models/manageiq/providers/openstack/infra_manager/refresher_rhos_juno_spec.rb
@@ -12,67 +12,49 @@ describe ManageIQ::Providers::Openstack::InfraManager::Refresher do
       :default => {:userid => "admin", :password => "a5d6375470291c68de726836504d014ebe095b6d"})
   end
 
-  it "graph refresh the same as classic refresh" do
-    stub_settings_merge(:ems_refresh => {:openstack_infra => {:inventory_object_refresh => false}})
-    full_refresh
-    before = serialize_inventory
+  it "will perform a full refresh" do
+    2.times do # Run twice to verify that a second run with existing data does not change anything
+      full_refresh
 
-    stub_settings_merge(:ems_refresh => {:openstack_infra => {:inventory_object_refresh => true}})
-    full_refresh
-    after = serialize_inventory
-
-    assert_inventory_not_changed(before, after)
+      assert_table_counts
+      assert_ems
+      assert_specific_host
+      assert_mapped_stacks
+      assert_specific_public_template
+    end
   end
 
-  [{:inventory_object_refresh => false}, {:inventory_object_refresh => true}].each do |refresh_settings|
-    context "with refresh settings #{refresh_settings}" do
-      before { stub_settings_merge(:ems_refresh => {:openstack_infra => refresh_settings}) }
+  it "will verify maintenance mode" do
+    # We need VCR to match requests differently here because fog adds a dynamic
+    #   query param to avoid HTTP caching - ignore_awful_caching##########
+    #   https://github.com/fog/fog/blob/master/lib/fog/openstack/compute.rb#L308
+    VCR.use_cassette("#{described_class.name.underscore}_rhos_juno_maintenance",
+                     :match_requests_on => [:method, :host, :path]) do
+      @ems.reload
+      @ems.reset_openstack_handle
+      Fog::OpenStack.instance_variable_set(:@version, nil)
+      EmsRefresh.refresh(@ems)
+      EmsRefresh.refresh(@ems.network_manager)
+      @ems.reload
 
-      it "will perform a full refresh" do
-        2.times do # Run twice to verify that a second run with existing data does not change anything
-          full_refresh
+      @host = ManageIQ::Providers::Openstack::InfraManager::Host.all.order(:ems_ref).detect { |x| x.name.include?('(NovaCompute)') }
 
-          assert_table_counts
-          assert_ems
-          assert_specific_host
-          assert_mapped_stacks
-          assert_specific_public_template
-        end
-      end
+      expect(@host.maintenance).to eq(false)
+      expect(@host.maintenance_reason).to be nil
 
-      it "will verify maintenance mode" do
-        # We need VCR to match requests differently here because fog adds a dynamic
-        #   query param to avoid HTTP caching - ignore_awful_caching##########
-        #   https://github.com/fog/fog/blob/master/lib/fog/openstack/compute.rb#L308
-        VCR.use_cassette("#{described_class.name.underscore}_rhos_juno_maintenance",
-                         :match_requests_on => [:method, :host, :path]) do
-          @ems.reload
-          @ems.reset_openstack_handle
-          Fog::OpenStack.instance_variable_set(:@version, nil)
-          EmsRefresh.refresh(@ems)
-          EmsRefresh.refresh(@ems.network_manager)
-          @ems.reload
+      @host.set_node_maintenance
+      EmsRefresh.refresh(@ems)
+      @ems.reload
+      @host.reload
+      expect(@host.maintenance).to eq(true)
+      expect(@host.maintenance_reason).to eq("CFscaledown")
 
-          @host = ManageIQ::Providers::Openstack::InfraManager::Host.all.order(:ems_ref).detect { |x| x.name.include?('(NovaCompute)') }
-
-          expect(@host.maintenance).to eq(false)
-          expect(@host.maintenance_reason).to be nil
-
-          @host.set_node_maintenance
-          EmsRefresh.refresh(@ems)
-          @ems.reload
-          @host.reload
-          expect(@host.maintenance).to eq(true)
-          expect(@host.maintenance_reason).to eq("CFscaledown")
-
-          @host.unset_node_maintenance
-          EmsRefresh.refresh(@ems)
-          @ems.reload
-          @host.reload
-          expect(@host.maintenance).to eq(false)
-          expect(@host.maintenance_reason).to be nil
-        end
-      end
+      @host.unset_node_maintenance
+      EmsRefresh.refresh(@ems)
+      @ems.reload
+      @host.reload
+      expect(@host.maintenance).to eq(false)
+      expect(@host.maintenance_reason).to be nil
     end
   end
